# Improving deletes performance by limiting number of records to scan