--- a/core/chains/evm/logpoller/orm_test.go
+++ b/core/chains/evm/logpoller/orm_test.go
@@ -1435,7 +1435,7 @@ func TestInsertLogsInTx(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			// clean all logs and blocks between test cases
-			defer func() { _, _ = db.NamedExec("truncate evm.logs", struct{}{}) }()
+			defer func() { _, _ = db.Exec("truncate evm.logs") }()
 
 			insertErr := o.InsertLogs(tt.logs)
 			logsFromDb, err := o.SelectLogs(0, math.MaxInt, address, event)