@reductionista 2024-02-13
I'm really surprised that adding the bounds made a difference in the plan, as the planner is supposed to consult the `pg_stats` table which includes a histogram of values in each table.  In principle, it should know what the range of values is in the table... I can't argue with the EXPLAIN ANALYZE though, it clearly makes a difference somehow.